Move some notebook-specific stuff from shared to notebookgenerator.
